Fix teleop review feedback
